resume builder/requirements.txt
Flask>=3.0
Flask-Login>=0.6
Flask-SQLAlchemy>=3.1
Werkzeug>=3.0
itsdangerous>=2.2
python-docx>=1.1.0
xhtml2pdf>=0.2.15
reportlab>=4.2.2
bleach>=6.1.0
pillow>=10.3.0
beautifulsoup4>=4.12.3
